backend/requirements.txt
fastapi
uvicorn
rembg
pillow
python-multipart
onnxruntime
opencv-python
numpy
realesrgan
basicsr
gfpgan
